requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.32.0
mtcnn>=0.1.1
Pillow>=9.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
lz4==4.3.3
joblib==1.3.2
